refactor: Extract menu from app controller file
