+++ packages/neuron-wallet/src/controllers/app/index.ts
@@ -2,17 +2,13 @@ import path from 'path'
 import {
   dialog,
   shell,
-  Menu,
   MenuItem,
-  MenuItemConstructorOptions,
-  clipboard,
   MessageBoxOptions,
   MessageBoxReturnValue,
   BrowserWindow,
 } from 'electron'
 import windowStateKeeper from 'electron-window-state'
 import { take } from 'rxjs/operators'
-import { bech32Address, AddressPrefix, AddressType } from '@nervosnetwork/ckb-sdk-utils'
 
 import app from 'app'
 import { TransactionsController, WalletsController, SyncInfoController, UpdateController } from 'controllers'
@@ -29,470 +25,13 @@ import { ConnectionStatusSubject } from 'models/subjects/node'
 import { SystemScriptSubject } from 'models/subjects/system-script'
 
 import { subscribe } from './subscribe'
-
-const isMac = process.platform === 'darwin'
-const networksService = NetworksService.getInstance()
+import { updateApplicationMenu, popContextMenu } from './menu'
 
 enum URL {
   Preference = '/settings/general',
   CreateWallet = '/wizard/mnemonic/create',
   ImportMnemonic = '/wizard/mnemonic/import',
   ImportKeystore = '/keystore/import',
-}
-
-enum ExternalURL {
-  Website = 'https://www.nervos.org/',
-  Repository = 'https://github.com/nervosnetwork/neuron',
-  Issues = 'https://github.com/nervosnetwork/neuron/issues',
-}
-
-const separator: MenuItemConstructorOptions = {
-  type: 'separator',
-}
-
-const generateTemplate = () => {
-  const appMenuItem: MenuItemConstructorOptions = {
-    id: 'app',
-    label: app.getName(),
-    submenu: [
-      {
-        id: 'about',
-        label: i18n.t('application-menu.neuron.about', {
-          app: app.getName(),
-        }),
-        role: 'about',
-        click: () => {
-          if (AppController) {
-            AppController.showAbout()
-          }
-        },
-      },
-      {
-        label: i18n.t('application-menu.neuron.check-updates'),
-        click: (menuItem: MenuItem) => {
-          if (AppController) {
-            AppController.checkUpdates(menuItem)
-          }
-        },
-      },
-      separator,
-      {
-        id: 'preference',
-        label: i18n.t('application-menu.neuron.preferences'),
-        accelerator: 'CmdOrCtrl+,',
-        click: () => {
-          if (AppController) {
-            AppController.showPreference()
-          }
-        },
-      },
-      separator,
-      {
-        label: i18n.t('application-menu.neuron.quit', {
-          app: app.getName(),
-        }),
-        role: 'quit',
-      },
-    ],
-  }
-
-  const walletMenuItem: MenuItemConstructorOptions = {
-    id: 'wallet',
-    label: i18n.t('application-menu.wallet.label'),
-    submenu: [
-      { id: 'select', label: i18n.t('application-menu.wallet.select'), submenu: [] },
-      {
-        id: 'create',
-        label: i18n.t('application-menu.wallet.create-new'),
-        click: () => {
-          if (AppController) {
-            AppController.createWallet()
-          }
-        },
-      },
-      {
-        id: 'import',
-        label: i18n.t('application-menu.wallet.import'),
-        submenu: [
-          {
-            id: 'import-with-mnemonic',
-            label: i18n.t('application-menu.wallet.import-mnemonic'),
-            click: () => {
-              if (AppController) {
-                AppController.importWallet('mnemonic')
-              }
-            },
-          },
-          {
-            id: 'import-with-keystore',
-            label: i18n.t('application-menu.wallet.import-keystore'),
-            click: () => {
-              if (AppController) {
-                AppController.importWallet('keystore')
-              }
-            },
-          },
-        ],
-      },
-      separator,
-      {
-        id: 'backup',
-        label: i18n.t('application-menu.wallet.backup'),
-        click: () => {
-          const walletsService = WalletsService.getInstance()
-          const currentWallet = walletsService.getCurrent()
-          if (!currentWallet) {
-            // TODO: show the error message
-            return
-          }
-          walletsService.requestPassword(currentWallet.id, 'backup-wallet')
-        },
-      },
-      {
-        id: 'delete',
-        label: i18n.t('application-menu.wallet.delete'),
-        click: () => {
-          const walletsService = WalletsService.getInstance()
-          const currentWallet = walletsService.getCurrent()
-          if (!currentWallet) {
-            // TODO: show the error message
-            return
-          }
-          walletsService.requestPassword(currentWallet.id, 'delete-wallet')
-        },
-      },
-    ],
-  }
-
-  const editMenuItem: MenuItemConstructorOptions = {
-    id: 'edit',
-    label: i18n.t('application-menu.edit.label'),
-    submenu: [
-      {
-        label: i18n.t('application-menu.edit.cut'),
-        role: 'cut',
-      },
-      {
-        label: i18n.t('application-menu.edit.copy'),
-        role: 'copy',
-      },
-      {
-        label: i18n.t('application-menu.edit.paste'),
-        role: 'paste',
-      },
-      separator,
-      {
-        label: i18n.t('application-menu.edit.selectall'),
-        role: 'selectAll',
-      },
-    ],
-  }
-
-  const viewMenuItem: MenuItemConstructorOptions = {
-    id: 'view',
-    label: i18n.t('application-menu.view.label'),
-    submenu: [
-      {
-        label: i18n.t('application-menu.view.fullscreen'),
-        role: 'togglefullscreen',
-      },
-      {
-        label: i18n.t('application-menu.view.address-book'),
-        click: () => {
-          if (AppController) {
-            AppController.toggleAddressBook()
-          }
-        },
-        accelerator: 'CmdOrCtrl+B',
-      },
-    ],
-  }
-
-  const windowMenuItem: MenuItemConstructorOptions = {
-    id: 'window',
-    label: i18n.t('application-menu.window.label'),
-    submenu: [
-      {
-        label: i18n.t('application-menu.window.minimize'),
-        role: 'minimize',
-      },
-      {
-        label: i18n.t('application-menu.window.close'),
-        role: 'close',
-      },
-    ],
-  }
-
-  const helpSubmenu: MenuItemConstructorOptions[] = [
-    {
-      label: 'Nervos',
-      click: () => {
-        if (AppController) {
-          AppController.openExternal(ExternalURL.Website)
-        }
-      },
-    },
-    {
-      label: i18n.t('application-menu.help.source-code'),
-      click: () => {
-        if (AppController) {
-          AppController.openExternal(ExternalURL.Repository)
-        }
-      },
-    },
-    {
-      label: i18n.t('application-menu.help.report-issue'),
-      click: () => {
-        if (AppController) {
-          AppController.openExternal(ExternalURL.Issues)
-        }
-      },
-    },
-  ]
-  if (!isMac) {
-    helpSubmenu.push(separator)
-    helpSubmenu.push({
-      id: 'preference',
-      label: i18n.t('application-menu.help.settings'),
-      click: () => {
-        if (AppController) {
-          AppController.showPreference()
-        }
-      },
-    })
-    helpSubmenu.push({
-      label: i18n.t('application-menu.neuron.check-updates'),
-      click: (menuItem: MenuItem) => {
-        if (AppController) {
-          AppController.checkUpdates(menuItem)
-        }
-      },
-    })
-    helpSubmenu.push({
-      id: 'about',
-      label: i18n.t('application-menu.neuron.about', {
-        app: app.getName(),
-      }),
-      role: 'about',
-      click: () => {
-        if (AppController) {
-          AppController.showAbout()
-        }
-      },
-    })
-  }
-
-  const helpMenuItem: MenuItemConstructorOptions = {
-    id: 'help',
-    label: i18n.t('application-menu.help.label'),
-    role: 'help',
-    submenu: helpSubmenu,
-  }
-
-  const developMenuItem: MenuItemConstructorOptions = {
-    id: 'develop',
-    label: i18n.t('application-menu.develop.develop'),
-    submenu: [
-      {
-        label: i18n.t('application-menu.develop.reload'),
-        role: 'reload',
-      },
-      {
-        label: i18n.t('application-menu.develop.force-reload'),
-        role: 'forceReload',
-      },
-      {
-        label: i18n.t('application-menu.develop.toggle-dev-tools'),
-        role: 'toggleDevTools',
-      },
-    ],
-  }
-
-  const applicationMenuTemplate = env.isDevMode
-    ? [walletMenuItem, editMenuItem, viewMenuItem, developMenuItem, windowMenuItem, helpMenuItem]
-    : [walletMenuItem, editMenuItem, viewMenuItem, windowMenuItem, helpMenuItem]
-
-  if (isMac) {
-    applicationMenuTemplate.unshift(appMenuItem)
-  }
-  return applicationMenuTemplate
-}
-
-const contextMenuTemplate: {
-  [key: string]: (id: string) => Promise<MenuItemConstructorOptions[]>
-} = {
-  copyMainnetAddress: async (identifier: string) => {
-    const address = bech32Address(identifier, {
-      prefix: AddressPrefix.Mainnet,
-      type: AddressType.HashIdx,
-      codeHashIndex: '0x00',
-    })
-    return [
-      {
-        label: i18n.t('contextMenu.copy-address'),
-        click: () => {
-          clipboard.writeText(address)
-        },
-      },
-    ]
-  },
-  networkList: async (id: string) => {
-    const [network, currentNetworkID] = await Promise.all([
-      networksService.get(id).catch(() => null),
-      networksService.getCurrentID().catch(() => null),
-    ])
-
-    if (!network) {
-      AppController.showMessageBox({
-        type: 'error',
-        message: i18n.t('messages.network-not-found', { id }),
-      })
-      return []
-    }
-
-    const isCurrent = currentNetworkID === id
-    const isDefault = network.type === 0
-
-    return [
-      {
-        label: i18n.t('contextMenu.select'),
-        enabled: !isCurrent,
-        click: () => {
-          networksService.activate(id).catch((err: Error) => {
-            AppController.showMessageBox({
-              type: 'error',
-              message: err.message,
-            })
-          })
-        },
-      },
-      {
-        label: i18n.t('contextMenu.edit'),
-        enabled: !isDefault,
-        click: () => {
-          AppController.navTo(`/network/${id}`)
-        },
-      },
-      {
-        label: i18n.t('contextMenu.delete'),
-        enabled: !isDefault,
-        cancelId: 1,
-        click: async () => {
-          AppController.showMessageBox(
-            {
-              type: 'warning',
-              title: i18n.t(`messageBox.remove-network.title`),
-              message: i18n.t(`messageBox.remove-network.message`, {
-                name: network.name,
-                address: network.remote,
-              }),
-              detail: isCurrent ? i18n.t('messageBox.remove-network.alert') : '',
-              buttons: [i18n.t('messageBox.button.confirm'), i18n.t('messageBox.button.discard')],
-            },
-            (returnValue: MessageBoxReturnValue) => {
-              if (returnValue.response === 0) {
-                try {
-                  networksService.delete(id)
-                } catch (err) {
-                  dialog.showMessageBox({
-                    type: 'error',
-                    message: err.message,
-                  })
-                }
-              }
-            }
-          )
-        },
-      },
-    ]
-  },
-  walletList: async (id: string) => {
-    const walletsService = WalletsService.getInstance()
-    const wallet = walletsService.get(id)
-    if (!wallet) {
-      AppController.showMessageBox({
-        type: 'error',
-        message: i18n.t('messages.wallet-not-found', { id }),
-      })
-    }
-    return [
-      {
-        label: i18n.t('contextMenu.select'),
-        click: () => {
-          try {
-            walletsService.setCurrent(id)
-          } catch (err) {
-            AppController.showMessageBox({ type: 'error', message: err.message })
-          }
-        },
-      },
-      {
-        label: i18n.t('contextMenu.backup'),
-        click: async () => {
-          walletsService.requestPassword(id, 'backup-wallet')
-        },
-      },
-      {
-        label: i18n.t('contextMenu.edit'),
-        click: () => {
-          AppController.navTo(`/editwallet/${id}`)
-        },
-      },
-      {
-        label: i18n.t('contextMenu.delete'),
-        click: async () => {
-          walletsService.requestPassword(id, 'delete-wallet')
-        },
-      },
-    ]
-  },
-  addressList: async (identifier: string) => {
-    if (identifier === undefined) {
-      return []
-    }
-
-    const address = bech32Address(identifier)
-    return [
-      {
-        label: i18n.t('contextMenu.copy-address'),
-        click: () => {
-          clipboard.writeText(address)
-        },
-      },
-      {
-        label: i18n.t('contextMenu.request-payment'),
-        click: () => {
-          AppController.navTo(`/receive/${address}`)
-        },
-      },
-      {
-        label: i18n.t('contextMenu.view-on-explorer'),
-        click: () => {
-          AppController.openExternal(`${env.explorer}/address/${address}`)
-        },
-      },
-    ]
-  },
-  transactionList: async (hash: string) => {
-    return [
-      {
-        label: i18n.t('contextMenu.detail'),
-        click: () => AppController.showTransactionDetails(hash),
-      },
-      {
-        label: i18n.t('contextMenu.copy-transaction-hash'),
-        click: () => {
-          clipboard.writeText(hash)
-        },
-      },
-      {
-        label: i18n.t('contextMenu.view-on-explorer'),
-        click: () => {
-          AppController.openExternal(`${env.explorer}/transaction/${hash}`)
-        },
-      },
-    ]
-  },
 }
 
 // Acts as a middle man so that subscribe doesn't have to know AppController
@@ -671,25 +210,11 @@ export default class AppController {
   }
 
   public static updateApplicationMenu = (wallets: Controller.Wallet[], id: string | null) => {
-    const applicationMenu = Menu.buildFromTemplate(generateTemplate())
-    const selectMenu = applicationMenu.getMenuItemById('select')
+    updateApplicationMenu(wallets, id)
+  }
 
-    wallets.forEach(wallet => {
-      selectMenu.submenu.append(
-        new MenuItem({
-          id: wallet.id,
-          label: wallet.name,
-          type: 'radio',
-          checked: wallet.id === id,
-          click: () => {
-            const walletsService = WalletsService.getInstance()
-            walletsService.setCurrent(wallet.id)
-          },
-        })
-      )
-    })
-
-    Menu.setApplicationMenu(applicationMenu)
+  public static async contextMenu(params: { type: string; id: string }) {
+    return popContextMenu(params)
   }
 
   public static toggleAddressBook() {
@@ -710,27 +235,6 @@ export default class AppController {
 
   public static openExternal(url: string) {
     shell.openExternal(url)
-  }
-
-  public static async contextMenu(params: { type: string; id: string }) {
-    if (!params || params.id === undefined) {
-      return
-    }
-    const { id, type } = params
-    switch (type) {
-      case 'copyMainnetAddress':
-      case 'networkList':
-      case 'walletList':
-      case 'addressList':
-      case 'transactionList': {
-        const menu = Menu.buildFromTemplate(await contextMenuTemplate[type](id))
-        menu.popup()
-        break
-      }
-      default: {
-        break
-      }
-    }
   }
 
   public static showAbout() {
